object2kafka/requirements.txt
fdk>=0.1.35
oci
kafka-python